Fix part of the failing tests #543
