+++ scanpipe/tests/test_api.py
@@ -651,7 +651,7 @@ class ScanPipeAPITest(TransactionTestCase):
             get_model_serializer(None)
 
     def test_scanpipe_api_serializer_get_serializer_fields(self):
-        self.assertEqual(30, len(get_serializer_fields(DiscoveredPackage)))
+        self.assertEqual(38, len(get_serializer_fields(DiscoveredPackage)))
         self.assertEqual(11, len(get_serializer_fields(DiscoveredDependency)))
         self.assertEqual(28, len(get_serializer_fields(CodebaseResource)))
